fix test and rename datetime helper to prevent import errors
