+++ src/dataland_qa_lab/pages/text_to_doc_intelligence.py
@@ -1,5 +1,3 @@
-from datetime import UTC, datetime, timedelta, timezone
-
 import pypdf
 from azure.ai.documentintelligence import DocumentIntelligenceClient
 from azure.ai.documentintelligence.models import DocumentContentFormat
@@ -8,6 +6,7 @@ from azure.core.credentials import AzureKeyCredential
 from dataland_qa_lab.database.database_engine import add_entity, get_entity
 from dataland_qa_lab.database.database_tables import ReviewedDatasetMarkdowns
 from dataland_qa_lab.utils import config
+from dataland_qa_lab.utils.datetime_helper import get_german_time_as_string
 
 
 def extract_text_of_pdf(pdf: pypdf.PdfReader) -> str:
@@ -31,9 +30,7 @@ def get_markdown_from_dataset(data_id: str, relevant_pages_pdf_reader: pypdf.Pdf
     """Adds or updates a markdown document in the database if necessary and returns it."""
     readable_text = None
 
-    now_utc = datetime.now(UTC)
-    ger_timezone = timedelta(hours=2) if now_utc.astimezone(timezone(timedelta(hours=1))).dst() else timedelta(hours=1)
-    formatted_german_time = (now_utc + ger_timezone).strftime("%Y-%m-%d %H:%M:%S")
+    german_time = get_german_time_as_string()
 
     exist_markdown = get_entity(entity_id=data_id, entity_class=ReviewedDatasetMarkdowns)
 
@@ -49,8 +46,8 @@ def get_markdown_from_dataset(data_id: str, relevant_pages_pdf_reader: pypdf.Pdf
             data_id=data_id,
             markdown_text=readable_text,
             page_numbers=page_numbers,
-            last_saved=formatted_german_time,
-            last_updated=formatted_german_time,
+            last_saved=german_time,
+            last_updated=german_time,
         )
 
         add_entity(new_document)
