+++ tests/database/test_text_to_doc_intelligence.py
@@ -34,14 +34,11 @@ def test_extract_text_of_pdf(mock_credential: MagicMock, mock_client: MagicMock,
 @patch("dataland_qa_lab.pages.text_to_doc_intelligence.extract_text_of_pdf")
 @patch("dataland_qa_lab.pages.text_to_doc_intelligence.add_entity")
 @patch("dataland_qa_lab.pages.text_to_doc_intelligence.get_entity")
-@patch("dataland_qa_lab.pages.text_to_doc_intelligence.create_tables")
 def test_get_markdown_from_dataset_new_entry(
-    mock_create_tables: MagicMock,
     mock_get_entity: MagicMock,
     mock_add_entity: MagicMock,
     mock_extract_text: MagicMock,
 ) -> None:
-    mock_create_tables.return_value = None
     mock_extract_text.return_value = "mocked_text"
     mock_get_entity.return_value = None
 
@@ -51,7 +48,6 @@ def test_get_markdown_from_dataset_new_entry(
 
     result = get_markdown_from_dataset(data_id=data_id, relevant_pages_pdf_reader=pdf_reader, page_numbers=pages)
 
-    mock_create_tables.assert_called_once()
     mock_get_entity.assert_called_once()
     mock_extract_text.assert_called_once_with(pdf_reader)
     mock_add_entity.assert_called_once()
@@ -60,14 +56,10 @@ def test_get_markdown_from_dataset_new_entry(
 
 @patch("dataland_qa_lab.pages.text_to_doc_intelligence.add_entity")
 @patch("dataland_qa_lab.pages.text_to_doc_intelligence.get_entity")
-@patch("dataland_qa_lab.pages.text_to_doc_intelligence.create_tables")
 def test_get_markdown_from_dataset_existing_entry(
-    mock_create_tables: MagicMock,
     mock_get_entity: MagicMock,
     mock_add_entity: MagicMock,
 ) -> None:
-    mock_create_tables.return_value = None
-
     mock_existing_entry = MagicMock()
     mock_existing_entry.pages = [1, 2]
     mock_existing_entry.markdown_text = "old_text"
@@ -80,7 +72,6 @@ def test_get_markdown_from_dataset_existing_entry(
 
     result = get_markdown_from_dataset(data_id=data_id, relevant_pages_pdf_reader=pdf_reader, page_numbers=pages)
 
-    mock_create_tables.assert_called_once()
     mock_get_entity.assert_called_once()
     mock_add_entity.assert_not_called()
     assert result == "old_text"
